Fin de la pygamisation du projet avec l'utilisation complète de Pygame pour l'affichage et la gestion des éléments graphiques.
Changements majeurs :
- Introduction d’une nouvelle classe de base ElementSprite pour regrouper les attributs et méthodes communs à Tuile et Decoration.
- Mise à jour des classes Tuile et Decoration afin d’hériter de ElementSprite, simplifiant ainsi leur implémentation.
- Modification de la classe Carte pour gérer les tuiles via un groupe de sprites, améliorant le rendu et l’organisation.
- Ajout d’une classe Camera permettant de gérer le mouvement de la caméra et le rendu des sprites visibles.

Changements mineurs :
- Ajustement du type des positions des éléments graphiques (tuile: int, decoration: float) pour une meilleure lisibilité et cohérence.
- Harmonisation des types de paramètres pour une meilleure clarté et cohérence dans tout le code.
- Suppression du fichier en double et inutilisé scripts/principal/noyau/tuile.py au profit de scripts/tuile.py.
- La base de la gestion des décorations est légèrement posée, mais pas du tout implémentée ni intégrée dans le rendu.

+++ scripts/parametres/__init__.py
@@ -4,7 +4,7 @@ import pygame
 from scripts.type import TypeDecoration, TypeTuile
 
 #* Jeu en général
-TAILLE_ECRAN: tuple[float, float] = (1690, 800)
+TAILLE_ECRAN: tuple[int, int] = (1690, 800)
 TAILLE_AFFICHAGE: tuple[float, float] = (TAILLE_ECRAN[0] * 0.98, TAILLE_ECRAN[1] * 0.98)
 FPS: int = 60
 VITESSE_CAMERA: int = 10
